requirements.txt
blinker==1.4
click==7.1.2
Flask==1.1.2
Flask-Mail==0.9.1
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2-binary==2.8.6
Werkzeug==1.0.1
